forgetpwd

+++ app/View/Themed/Hoebll/Account/index.ctp
@@ -113,8 +113,8 @@
             <tr>
                 <td colspan="100">
                     <!-- Player related actions -->
-                    <a href="/account/forms" title="Fill out League Required forms">File Forms</a> | <a href="/account/editplayer/<?php echo $player['player_id'];?>" title="Edit this player">Edit</a> | 
-                    <a href="/account/deleteplayer" title="Delete this player">Delete</a>
+                    <a href="/account/playerforms/<?php echo $player['player_id'];?>" title="Fill out League Required forms">File Forms</a> | <a href="/account/editplayer/<?php echo $player['player_id'];?>" title="Edit this player">Edit</a> | 
+                    <a href="/account/deleteplayer/<?php echo $player['player_id'];?>" title="Delete this player" onclick="return confirm('<?php echo __('Are you sure');?>?')">Delete</a>
                 </td>
             </tr>
             <?
